MainPaths.jl

Welcome to MainPaths.jl, a Julia package for computing main paths built on top of the LightGraphs.jl graph library.

Getting started

The package is currently not registered but can be added directly via the github package URL.

From the julia REPL, type ] to enter the Pkg REPL and run:

pkg> add https://github.com/jfb-h/MainPaths.jl

You can then run:

>julia using MainPaths

to get started.

Example

As a showcase of the package's functionality, we compute the forward-backward main path based on SPC weight sof a simple example graph:

First, define the graph:

using LightGraphs, MainPaths

A = [
 0  0  1  0  0  0  0  0  0  0  0
 0  0  1  1  0  0  0  0  0  0  0
 0  0  0  0  1  0  1  0  0  0  0
 0  0  0  0  0  1  0  0  0  1  1
 0  0  0  0  0  0  0  1  1  0  0
 0  0  0  0  0  0  0  0  1  1  0
 0  0  0  0  0  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0
];

julia> g = SimpleDiGraph(A)
{11, 12} directed simple Int64 graph

Then compute traversal weights:

julia> eweights, vweighs, flow = MainPaths.weights_spc(g, normalize=false);

And finally, compute the main path starting at nodes 1 and 2 (we here perform the main path traversal using edge weights, but vertex weights would also be possible):

julia> start = [1, 2];
julia> mp = mainpath(testgraph, start, eweights, FBMP())
MainPath with 8 vertices and 9 edges.

API

MainPaths.bfs_multi — Method

bfs_multi(g, source, neighborfn)

Traverse graph g via breadth-first search starting at vertices source and choosing neighbors to visit according to neighborfn.

This is a modified breadth-first search traversal where instances of a vertex being visited multiple times by different parents are recorded. This function was modified from the LightGraphs bfs algorithm.

MainPaths.emax_inneighbors — Method

emax_inneighbors(g, v, w)

Obtain an array with the outneighbor(s) of vertex v that are connected with the maximum weight edge in terms of edgeweight w.

MainPaths.emax_outneighbors — Method

emax_outneighbors(g, v, w)

Obtain an array with the outneighbor(s) of vertex v that are connected with the maximum weight edge in terms of edgeweight w.

MainPaths.mainpath — Method

mainpath(g, start, weights::Matrix{<:Real}, FBMP())

Compute the main path or main path network for graph g, traversing both forward and backward starting at nodes start and using edge weights given as matrix weights for neighbor selection during traversal.

MainPaths.mainpath — Method

mainpath(g, start, weights::Vector{<:Real}, FBMP())

Compute the main path or network of main paths for graph g, traversing both forward and backward starting at nodes start and using node weights given as vector weights for neighbor selection during traversal.

MainPaths.vmax_inneighbors — Method

vmax_inneighbors(g, v, w)

Obtain an array with the inneighbor(s) of vertex v that score highest in terms of vertex weights w.

MainPaths.vmax_outneighbors — Method

vmax_outneighbors(g, v, w)

Obtain an array with the outneighbor(s) of vertex v that score highest in terms of vertex weights w.

MainPaths.weights_gkp — Method

weights_gkp(g, normalize=:none)

Compute the genetic knowledge persistence scores for all nodes in g.

This is a modified version of the method proposed in Martinelli & Nomaler (2014) where citations from future patents to patents before the curent layer are not ignored in the persistence computation.

MainPaths.weights_spc — Method

weights_spc(g, normalize=false)

Compute the Search Path Count (SPC) weights for graph g as described in Batagelj (2003). The function returns a vector with edgeweights, a vector with vertex weights and the total flow. When normalize is set to true, edge weigths and vertex weights are divided by total flow.

MainPaths.FBMP — Type

Struct representing a forward-backward main path traversal. This doesn't contain any information but is used for dispatch in the mainpath() methods.